Add content_hash to feed_items
Currently, for `feed_items` (feed_id, link, title) are used as primary
key.

combination of link and title for some feeds is too big for some feeds
and can not be used as primary key.

This pr is an attempt to use a generated hash field instead as part of
primary key